# Improve performance by using hstore to compute row_changed_fields