I noticed the performance hit when `hstore` was swapped out for `jsonb`.  I tracked it down to the query which calculated `changed_fields`, and found that be reverting to using the `hstore` `-` operator I was able to recover most of that performance.  It's still slower than `hstore`, but the gap is significantly shrunk.

No audit trigger: 9.7s
`hstore` original: 16.5s
`jsonb` original: 25.4s
`jsonb` This PR: 18.1s